chore: restructure project in workspaces
