+++ .yarnrc.yml
@@ -1,0 +1,2 @@
+# .yarnrc.yml
+nodeLinker: node-modules
