--- core/package.json
@@ -1,28 +1,0 @@
-{
-  "name": "@snapdrag/core",
-  "type": "module",
-  "version": "1.0.0",
-  "source": "src/index.ts",
-  "main": "./dist/snapdrag-core.cjs",
-  "module": "./dist/snapdrag-core.js",
-  "umd:main": "./dist/snapdrag-core.umd.cjs",
-  "unpkg": "./dist/snapdrag-core.umd.cjs",
-  "types": "./dist/types/index.d.ts",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/types/index.d.ts",
-        "default": "./dist/snapdrag-core.js"
-      },
-      "require": {
-        "types": "./dist/types/index.d.cts",
-        "default": "./dist/snapdrag-core.cjs"
-      }
-    }
-  },
-  "scripts": {
-    "build": "vite build --minify=false --sourcemap=true && cp ./dist/types/index.d.ts ./dist/types/index.d.cts"
-  },
-  "author": "Eugene Daragan",
-  "license": "MIT"
-}
