--- core/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "compilerOptions": {
-    "baseUrl": ".",
-    "strict": true,
-    "moduleResolution": "node",
-    "sourceMap": true,
-    "target": "ES2018",
-    "lib": ["ES2018", "DOM"]
-  },
-  "exclude": ["node_modules", "dist", "test"]
-}
