+++ packages/examples/.prettierrc
