+++ packages/examples/package.json
@@ -1,26 +1,27 @@
 {
-  "name": "snapdrag-test",
+  "name": "snapdrag-examples",
+  "version": "0.1.0",
   "private": true,
-  "version": "0.0.0",
   "type": "module",
   "scripts": {
-    "dev": "vite",
-    "build": "tsc && vite build",
-    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0",
-    "preview": "vite preview"
+    "build": "vite build",
+    "dev": "vite --open --host",
+    "lint": "eslint . --ext ts,tsx --report-unused-disable-directives --max-warnings 0 && yarn typecheck",
+    "preview": "vite preview",
+    "typecheck": "tsc --noEmit"
   },
   "dependencies": {
     "classnames": "^2.5.1",
     "lodash": "^4.17.21",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "snapdrag": "file:../",
+    "react": ">=18.3.0",
+    "react-dom": ">=18.3.0",
+    "snapdrag": "workspace:*",
     "styled-components": "^6.1.8"
   },
   "devDependencies": {
     "@types/lodash": "^4.17.1",
-    "@types/react": "^18.2.66",
-    "@types/react-dom": "^18.2.22",
+    "@types/react": ">=18.3.0",
+    "@types/react-dom": ">=18.3.0",
     "@typescript-eslint/eslint-plugin": "^7.2.0",
     "@typescript-eslint/parser": "^7.2.0",
     "@vitejs/plugin-react": "^4.2.1",
