+++ packages/examples/src/kanban-dashboard/dashboard.tsx
