+++ packages/examples/src/simple-squares/squares.tsx
