+++ packages/examples/src/simple-squares/styles.css
