+++ packages/examples/vite.config.ts
