+++ packages/snapdrag/package.json
@@ -1,0 +1,88 @@
+{
+  "name": "snapdrag",
+  "version": "0.8.0",
+  "description": "A simple, lightweight, and performant drag and drop library for React and vanilla JS",
+  "keywords": [
+    "drag",
+    "drop",
+    "drag-and-drop",
+    "draggable",
+    "vanilla",
+    "typescript",
+    "react",
+    "react-dnd"
+  ],
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/zheksoon/snapdrag.git"
+  },
+  "license": "MIT",
+  "author": {
+    "name": "Eugene Daragan"
+  },
+  "type": "module",
+  "exports": {
+    ".": {
+      "import": {
+        "types": "./dist/index.d.ts",
+        "default": "./dist/index.js"
+      },
+      "require": {
+        "types": "./dist/index.d.cts",
+        "default": "./dist/index.cjs"
+      }
+    },
+    "./core": {
+      "import": {
+        "types": "./dist/core.d.ts",
+        "default": "./dist/core.js"
+      },
+      "require": {
+        "types": "./dist/core.d.cts",
+        "default": "./dist/core.cjs"
+      }
+    },
+    "./plugins": {
+      "import": {
+        "types": "./dist/plugins.d.ts",
+        "default": "./dist/plugins.js"
+      },
+      "require": {
+        "types": "./dist/plugins.d.cts",
+        "default": "./dist/plugins.cjs"
+      }
+    }
+  },
+  "main": "./dist/index.cjs",
+  "module": "./dist/index.js",
+  "source": "./src/index.ts",
+  "types": "./dist/index.d.ts",
+  "files": [
+    "dist/**/*",
+    "src/**/*",
+    "package.json",
+    "README.md",
+    "LICENSE"
+  ],
+  "scripts": {
+    "build": "tsup --clean && yarn copy-files && yarn publint",
+    "copy-files": "cp ../../README.md ../../LICENSE dist",
+    "dev": "tsup --watch",
+    "lint": "yarn typecheck",
+    "typecheck": "tsc --noEmit"
+  },
+  "devDependencies": {
+    "@types/node": ">=20.18.0",
+    "@types/react": ">=18.3.0",
+    "publint": "^0.3.9",
+    "react": ">=18.3.0",
+    "tsup": "^8.4.0"
+  },
+  "peerDependencies": {
+    "react": ">=18.3.0"
+  },
+  "packageManager": "yarn@4.9.1",
+  "engines": {
+    "node": ">=20.18.0"
+  }
+}
