+++ packages/snapdrag/src/core/constants.ts
@@ -1,3 +1,2 @@
 export const DRAG_SOURCE_ATTRIBUTE = "data-drag-source";
-
 export const DROP_TARGET_ATTRIBUTE = "data-drop-target";
