--- react/package.json
@@ -1,36 +1,0 @@
-{
-  "name": "@snapdrag/react",
-  "type": "module",
-  "version": "1.0.0",
-  "source": "src/index.ts",
-  "main": "./dist/snapdrag-react.cjs",
-  "module": "./dist/snapdrag-react.js",
-  "umd:main": "./dist/snapdrag-react.umd.cjs",
-  "unpkg": "./dist/snapdrag-react.umd.cjs",
-  "types": "./dist/types/index.d.ts",
-  "exports": {
-    ".": {
-      "import": {
-        "types": "./dist/types/index.d.ts",
-        "default": "./dist/snapdrag-react.js"
-      },
-      "require": {
-        "types": "./dist/types/index.d.cts",
-        "default": "./dist/snapdrag-react.cjs"
-      }
-    }
-  },
-  "peerDependencies": {
-    "react": ">=16.8.0",
-    "snapdrag/core": "file:../core"
-  },
-  "devDependencies": {
-    "@types/react": ">=16.8.0",
-    "snapdrag/core": "file:../core"
-  },
-  "scripts": {
-    "build": "vite build --minify=false --sourcemap=true && cp dist/types/index.d.ts dist/types/index.d.cts"
-  },
-  "author": "Eugene Daragan",
-  "license": "MIT"
-}
